Support multiple types of date expressions
This is not needed for English, but in Finnish we need to support
multiple types of date expressions. Most importantly, it's not
sufficient to modify the expressions from a lemma with morphological
generation, as some changes also require the stem to change. For
example, while the English "today" is "tänään" in Finnish, "by today"
would be "tähän päivään mennessä", or literally "by this day". There is
not reasonable lemma which can be used in both cases.